tools/ai-markmap-agent/requirements.txt
# =============================================================================
# AI Markmap Agent - Dependencies
# =============================================================================
# Python >= 3.10 required
# =============================================================================

# -----------------------------------------------------------------------------
# Core: LangGraph & LangChain
# -----------------------------------------------------------------------------
langgraph>=1.0.4
langchain>=0.3.0
langchain-core>=0.3.0

# -----------------------------------------------------------------------------
# LLM Providers
# -----------------------------------------------------------------------------
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# -----------------------------------------------------------------------------
# Vector Store (Long-term Memory)
# -----------------------------------------------------------------------------
chromadb>=0.4.0
# pinecone-client>=3.0.0  # Uncomment if using Pinecone
# faiss-cpu>=1.7.4        # Uncomment if using FAISS

# -----------------------------------------------------------------------------
# Embeddings
# -----------------------------------------------------------------------------
tiktoken>=0.5.0

# -----------------------------------------------------------------------------
# Configuration & Utilities
# -----------------------------------------------------------------------------
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
tomli>=2.0.0;python_version<"3.11"  # TOML parser for Python < 3.11

# -----------------------------------------------------------------------------
# Markmap HTML Generation
# -----------------------------------------------------------------------------
jinja2>=3.1.0
markdown>=3.5.0

# -----------------------------------------------------------------------------
# CLI & Logging
# -----------------------------------------------------------------------------
click>=8.1.0
rich>=13.0.0
loguru>=0.7.0

# -----------------------------------------------------------------------------
# Development & Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0

# -----------------------------------------------------------------------------
# Optional: LangGraph Studio (Visualization)
# -----------------------------------------------------------------------------
# langgraph-studio>=0.1.0  # Uncomment if using LangGraph Studio

